Dev admin dashboard (#33)
* Temp commit for personal non-technical reasons. No meaningful changes

* Separate installation instructions from README. Include a bit more detail in installation.md

* UPdate environment vars for easy direction during installation

* Restructure javascript files to have them in folders depending on where they are used

* Initial implementation of admin dashboard. Doesnt include Js code yet.

* First draft of admin dashboard complete.

Visual structure of page present.
View functionality present (mostly).
Page built for desktop use only
Add functionality absent (cannot add users/modules/students)
Search functionality absent

* Allow admin login to admin landing page.

* minor clean up in landing page config and inclusion of a few minor todos

* Require login to see landing pages
