+++ nipo/static/javascript/staff_dashboard/markAttendance.js
