+++ nipo/static/javascript/student_dashboard/calcAttendance.js
