+++ nipo/templates/register.html
@@ -1,6 +1,7 @@
 {% extends "base.html" %}
 
 {% block content %}
+<!-- TODO: Change this page's styles to match the login page -->
     <h1>Register</h1>
     <form action="" method="post">
         {{ form.hidden_tag() }}
