+++ nipo/templates/staff_dashboard.html
@@ -56,6 +56,6 @@
 
 
 
-<script src="{{url_for('static',filename='javascript/markAttendance.js')}}"></script>
+<script src="{{url_for('static',filename='javascript/staff_dashboard/markAttendance.js')}}"></script>
 	
 {% endblock %}
