+++ nipo/templates/student_dashboard.html
@@ -41,6 +41,6 @@
 </div>
 
 
-<script src="{{url_for('static',filename='javascript/calcAttendance.js')}}"></script>
+<script src="{{url_for('static',filename='javascript/student_dashboard/calcAttendance.js')}}"></script>
 	
 {% endblock %}
