+++ requirements.txt
@@ -1,35 +1,55 @@
+appdirs==1.4.3
 atomicwrites==1.3.0
 attrs==19.1.0
+CacheControl==0.12.6
 certifi==2019.3.9
 chardet==3.0.4
 Click==7.0
+colorama==0.4.3
 configparser==3.7.4
+contextlib2==0.6.0
+distlib==0.3.0
+distro==1.4.0
 dlib==19.16.0
 docopt==0.6.2
 face-recognition==1.2.3
 face-recognition-models==0.3.0
 Flask==1.0.2
 Flask-Login==0.4.1
-Flask-WTF==0.14.2
+Flask-WTF==0.14.3
 grip==4.5.2
+html5lib==1.0.1
 idna==2.8
+ipaddr==2.2.0
 itsdangerous==1.1.0
 Jinja2==2.10.1
+lockfile==0.12.2
 Markdown==3.1.1
 MarkupSafe==1.1.1
-numpy==1.16.1
-opencv-python==4.0.0.21
+more-itertools==8.7.0
+msgpack==0.6.2
+numpy==1.20.1
+opencv-python==4.5.1.48
+packaging==20.3
+pandas==1.2.3
 path-and-address==2.0.1
+pep517==0.8.2
 Pillow==7.0.0
 pluggy==0.9.0
-psycopg2==2.7.7
-psycopg2-binary==2.7.7
+progress==1.5
+psycopg2-binary==2.8.6
 py==1.8.0
 Pygments==2.4.2
+pyparsing==2.4.6
 pytest==4.3.0
+python-dateutil==2.8.1
+pytoml==0.1.21
+pytz==2021.1
 requests==2.22.0
+retrying==1.3.3
 six==1.12.0
-SQLAlchemy==1.3.13
+SQLAlchemy==1.3.23
 urllib3==1.25.3
-Werkzeug==0.15.4
+webencodings==0.5.1
+Werkzeug==1.0.1
 WTForms==2.2.1
